+++ Sources/ParseSwift/Authentication/Protocols/ParseAuthentication.swift
@@ -103,7 +103,6 @@ public protocol ParseAuthentication: Codable {
      - parameter callbackQueue: The queue to return to after completion. Default value of .main.
      - parameter completion: The block to execute.
      */
-    @available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
     func loginPublisher(authData: [String: String],
                         options: API.Options) -> Future<AuthenticatedUser, ParseError>
 
@@ -114,7 +113,6 @@ public protocol ParseAuthentication: Codable {
      - parameter callbackQueue: The queue to return to after completion. Default value of .main.
      - parameter completion: The block to execute.
      */
-    @available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
     func linkPublisher(authData: [String: String],
                        options: API.Options) -> Future<AuthenticatedUser, ParseError>
 
@@ -126,7 +124,6 @@ public protocol ParseAuthentication: Codable {
      - parameter completion: The block to execute.
      It should have the following argument signature: `(Result<Self, ParseError>)`.
      */
-    @available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
     func unlinkPublisher(_ user: AuthenticatedUser,
                          options: API.Options) -> Future<AuthenticatedUser, ParseError>
 
@@ -137,7 +134,6 @@ public protocol ParseAuthentication: Codable {
      - parameter completion: The block to execute.
      It should have the following argument signature: `(Result<Self, ParseError>)`.
      */
-    @available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
     func unlinkPublisher(options: API.Options) -> Future<AuthenticatedUser, ParseError>
     #endif
 }