+++ Sources/ParseSwift/Protocols/CloudObservable.swift
@@ -12,7 +12,6 @@ import Foundation
  This protocol describes the interface for creating a view model for `ParseCloud` functions and jobs.
  You can use this protocol on any custom class of yours, instead of `CloudViewModel`, if it fits your use case better.
  */
-@available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
 public protocol CloudObservable: ObservableObject {
 
     /// The `ParseObject` associated with this view model.