+++ Sources/ParseSwift/Types/CloudViewModel.swift
@@ -12,7 +12,6 @@ import Foundation
  A default implementation of the `CloudCodeObservable` protocol. Suitable for `ObjectObserved`
  and can be used as a SwiftUI view model.
  */
-@available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
 open class CloudViewModel<T: ParseCloud>: CloudObservable {
 
     public typealias CloudCodeType = T
@@ -68,7 +67,6 @@ open class CloudViewModel<T: ParseCloud>: CloudObservable {
 }
 
 // MARK: CloudCodeViewModel
-@available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
 public extension ParseCloud {
 
     /**