+++ Sources/ParseSwift/Types/ParseAnalytics+combine.swift
@@ -14,7 +14,6 @@ import Combine
 import UIKit
 #endif
 
-@available(macOS 10.15, iOS 13.0, macCatalyst 13.0, watchOS 6.0, tvOS 13.0, *)
 public extension ParseAnalytics {
 
     // MARK: Combine